+++ src/services/masterportfolio/masterportfolio.ts
@@ -13,29 +13,10 @@ export const fetchMasterPortfolioList = async (cursor?: string) => {
   return res.data.result;
 };
 
-// 마스터 포트폴리오 목록을 가져오는 훅(마이페이지)
-export const useMasterPortfolioList = (cursor?: string) => {
-  return useQuery({
-    queryKey: ['master-portfolios', cursor],
-    queryFn: () => fetchMasterPortfolioList(cursor),
-    staleTime: 1000 * 60 * 5,
-  });
-};
-
 // 마스터 포트폴리오 상세 정보를 가져오는 함수(마스터포트폴리오)
 export const fetchMasterPortfolioDetail = async (projectId: number) => {
   const res: AxiosResponse<MasterPortfolioDetailResponse> = await axiosInstance.get(
     `/api/v1/master-portfolios/${projectId}`
   );
   return res.data.result;
-};
-
-// 마스터 포트폴리오 상세 정보를 가져오는 훅(마스터포트폴리오)
-export const useMasterPortfolioDetail = (projectId: number) => {
-  return useQuery({
-    queryKey: ['master-portfolio', projectId],
-    queryFn: () => fetchMasterPortfolioDetail(projectId),
-    staleTime: 1000 * 60 * 5,
-    enabled: !!projectId,
-  });
 };
